Add translation & encoding
Trades the #greet method for a #start method, which now initializes
translation functionality in addition to greeting the user.

Calling `OlaMundo.start` does this by including a group of methods in
the String class, like #translate, so that they can be called directly
on string objects. For example, `"Hello world!".translate("maltese")`
now translates English to Maltese.

This translate method transforms text to either binary (which defaults
to UTF-8 encoding in MSbit format) or the desired language depending on
the objective and whether or not it is currently available. Unknown
words and languages are encoded to binary instead.

Examples:
```
OlaMundo.start
=> "Olá mundo!"
"Hello world!".translate("maltese")
=> Bongu dinja!
"hello quatervois?".translate("italian")
=> "Ciao 01110001 01110101 01100001 01110100 01100101 01110010 01110110 01101111 01101001 01110011?"
"hi".translate("binary")
=> "01101000 01101001"
```

Adds a Transformer to hold these transformative methods that
are attached to strings, including #translate. Also adds an Encoder to
handle binary encoding, a Translator to map the full content of text
from one language to another, and a SectionTranslator to translate each
section of that content.